+++ pkg/collector/corechecks/snmp/internal/profile/profile_yaml.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 
@@ -198,51 +197,4 @@ func resolveProfileDefinitionPath(definitionFile string) string {
 func getProfileConfdRoot(profileFolderName string) string {
 	confdPath := config.Datadog.GetString("confd_path")
 	return filepath.Join(confdPath, "snmp.d", profileFolderName)
-}
-
-func getMostSpecificOid(oids []string) (string, error) {
-	var mostSpecificParts []int
-	var mostSpecificOid string
-
-	if len(oids) == 0 {
-		return "", fmt.Errorf("cannot get most specific oid from empty list of oids")
-	}
-
-	for _, oid := range oids {
-		parts, err := getOidPatternSpecificity(oid)
-		if err != nil {
-			return "", err
-		}
-		if len(parts) > len(mostSpecificParts) {
-			mostSpecificParts = parts
-			mostSpecificOid = oid
-			continue
-		}
-		if len(parts) == len(mostSpecificParts) {
-			for i := range mostSpecificParts {
-				if parts[i] > mostSpecificParts[i] {
-					mostSpecificParts = parts
-					mostSpecificOid = oid
-				}
-			}
-		}
-	}
-	return mostSpecificOid, nil
-}
-
-func getOidPatternSpecificity(pattern string) ([]int, error) {
-	wildcardKey := -1
-	var parts []int
-	for _, part := range strings.Split(strings.TrimLeft(pattern, "."), ".") {
-		if part == "*" {
-			parts = append(parts, wildcardKey)
-		} else {
-			intPart, err := strconv.Atoi(part)
-			if err != nil {
-				return nil, fmt.Errorf("error parsing part `%s` for pattern `%s`: %v", part, pattern, err)
-			}
-			parts = append(parts, intPart)
-		}
-	}
-	return parts, nil
 }
