+++ pkg/collector/corechecks/snmp/internal/profile/profile_yaml_test.go
@@ -51,65 +51,6 @@ func Test_getDefaultProfilesDefinitionFiles(t *testing.T) {
 	assert.Equal(t, expectedProfileConfig, actualProfileConfig)
 }
 
-func Test_getMostSpecificOid(t *testing.T) {
-	tests := []struct {
-		name           string
-		oids           []string
-		expectedOid    string
-		expectedErrror error
-	}{
-		{
-			"one",
-			[]string{"1.2.3.4"},
-			"1.2.3.4",
-			nil,
-		},
-		{
-			"error on empty oids",
-			[]string{},
-			"",
-			fmt.Errorf("cannot get most specific oid from empty list of oids"),
-		},
-		{
-			"error on parsing",
-			[]string{"a.1.2.3"},
-			"",
-			fmt.Errorf("error parsing part `a` for pattern `a.1.2.3`: strconv.Atoi: parsing \"a\": invalid syntax"),
-		},
-		{
-			"most lengthy",
-			[]string{"1.3.4", "1.3.4.1.2"},
-			"1.3.4.1.2",
-			nil,
-		},
-		{
-			"wild card 1",
-			[]string{"1.3.4.*", "1.3.4.1"},
-			"1.3.4.1",
-			nil,
-		},
-		{
-			"wild card 2",
-			[]string{"1.3.4.1", "1.3.4.*"},
-			"1.3.4.1",
-			nil,
-		},
-		{
-			"sample oids",
-			[]string{"1.3.6.1.4.1.3375.2.1.3.4.43", "1.3.6.1.4.1.8072.3.2.10"},
-			"1.3.6.1.4.1.3375.2.1.3.4.43",
-			nil,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			oid, err := getMostSpecificOid(tt.oids)
-			assert.Equal(t, tt.expectedErrror, err)
-			assert.Equal(t, tt.expectedOid, oid)
-		})
-	}
-}
-
 func Test_resolveProfileDefinitionPath(t *testing.T) {
 	defaultTestConfdPath, _ := filepath.Abs(filepath.Join("..", "test", "user_profiles.d"))
 	config.Datadog.Set("confd_path", defaultTestConfdPath)
