+++ pkg/collector/corechecks/snmp/internal/profile/sysobjectid_resolver.go
@@ -1,0 +1,59 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+package profile
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+)
+
+func getMostSpecificOid(oids []string) (string, error) {
+	var mostSpecificParts []int
+	var mostSpecificOid string
+
+	if len(oids) == 0 {
+		return "", fmt.Errorf("cannot get most specific oid from empty list of oids")
+	}
+
+	for _, oid := range oids {
+		parts, err := getOidPatternSpecificity(oid)
+		if err != nil {
+			return "", err
+		}
+		if len(parts) > len(mostSpecificParts) {
+			mostSpecificParts = parts
+			mostSpecificOid = oid
+			continue
+		}
+		if len(parts) == len(mostSpecificParts) {
+			for i := range mostSpecificParts {
+				if parts[i] > mostSpecificParts[i] {
+					mostSpecificParts = parts
+					mostSpecificOid = oid
+				}
+			}
+		}
+	}
+	return mostSpecificOid, nil
+}
+
+func getOidPatternSpecificity(pattern string) ([]int, error) {
+	wildcardKey := -1
+	var parts []int
+	for _, part := range strings.Split(strings.TrimLeft(pattern, "."), ".") {
+		if part == "*" {
+			parts = append(parts, wildcardKey)
+		} else {
+			intPart, err := strconv.Atoi(part)
+			if err != nil {
+				return nil, fmt.Errorf("error parsing part `%s` for pattern `%s`: %v", part, pattern, err)
+			}
+			parts = append(parts, intPart)
+		}
+	}
+	return parts, nil
+}
